Docs

+++ src/gameobject/Animal.java
@@ -1,5 +1,11 @@
 package gameobject;
 
+/**
+ * Animal is an abstract {@link GameObject} for all fish represented in our game.
+ * 
+ * @author W Mathieu Mimms-Boyce
+ * @author Miguel Fuentes
+ */
 public abstract class Animal extends GameObject {
 	private String name;
 	private int weight;
@@ -27,14 +33,35 @@ public abstract class Animal extends GameObject {
 	private String realPic = "";
 	private double avgSize = 0;
 
+	/**
+	 * Default constructor for the Animal class.
+	 * <p>
+	 * All parameters not set by this constructer will be set in the constructor
+	 * for each specific animal.
+	 * 
+	 * @see GameObject#GameObject(int, int, int, int, int, String) This constructor
+	 * sets everything the parent class does.
+	 * @see Animal#Animal(int, int, int, String, String, int, int, int, int, int, Question, int, int, boolean)
+	 * More specific debugging constructor
+	 */
 	public Animal(int xPos, int yPos, int depth, int xSize, int ySize) {
-		/*
-		 * This constructor will be used by children classes to use setters to assign
-		 * the other values which should be constants.
-		 */
 		super(xPos, yPos, depth, xSize, ySize, "");
 	}
 
+	/**
+	 * This should only be used for debugging purposes, as it is a very specific constructor.
+	 * 
+	 * @param name The name of the fish.
+	 * @param weight The weight of the fish (in lbs.)
+	 * @param size The size of the fish sprite.
+	 * @param speed The speed the fish moves at.
+	 * @param question The {@link Question} object for the fish.
+	 * @param pathState How far along the fish is on its pathfinding.
+	 * @param pathLength How long the fish's pathfinding is.
+	 * @param movingForward Whether the fish is moving forward or not.
+	 * @see Animal#Animal(int, int, int, int, int)
+	 * 
+	 */
 	private Animal(int xPos, int yPos, int depth, String imagePath, String name, int xSize, int ySize, int weight,
 			int size, int speed, Question question, int pathState, int pathLength, boolean movingForward) {
 		super(xPos, yPos, depth, xSize, ySize, imagePath);
@@ -57,9 +84,8 @@ public abstract class Animal extends GameObject {
 	}
 
 	/**
-	 * void updatePosition
-	 * 
-	 * this is used to advance through our pathfinding.
+	 * updatePositon is used to have fish move along their pathfinding.
+	 * This should be called in the update method of each fish.
 	 */
 	public void updatePosition() {
 		int advancement;
@@ -80,74 +106,128 @@ public abstract class Animal extends GameObject {
 		}
 	}
 
+	/**
+	 * @return The {@link Question} object of the fish.
+	 */
 	public Question getQuestion() {
 		return question;
 	}
 
+	/**
+	 * @param question Changes the {@link Question} object of the fish.
+	 */
 	public void setQuestion(Question question) {
 		this.question = question;
 	}
 
+	/**
+	 * @return The name of the fish.
+	 */
 	public String getName() {
 		return name;
 	}
 
+	/**
+	 * @param name What the fish will be renamed to.
+	 */
 	public void setName(String name) {
 		this.name = name;
 	}
 
+	/**
+	 * @return The weight of the fish.
+	 */
 	public int getWeight() {
 		return weight;
 	}
 
+	/**
+	 * @param weight The weight to set the fish to.
+	 */
 	public void setWeight(int weight) {
 		this.weight = weight;
 	}
 
+	/**
+	 * @return The speed of the fish.
+	 */
 	public int getSpeed() {
 		return speed;
 	}
 
+	/**
+	 * @param speed How fast the fish will be set to.
+	 */
 	public void setSpeed(int speed) {
 		this.speed = speed;
 	}
 
+	/**
+	 * @return The path length of the fish. This is how long the path is.
+	 */
 	public int getPathLength() {
 		return pathLength;
 	}
 
+	/**
+	 * @return The pathState of the fish. This is how far along the path the fish is.
+	 */
 	public int getPathState() {
 		return pathState;
 	}
 
+	/**
+	 * @param pathState What the pathState is going to be set to.
+	 */
 	public void setPathState(int pathState) {
 		this.pathState = pathState;
 	}
 
+	/**
+	 * @param pathLength What the pathLength is going to be set to.
+	 */
 	public void setPathLength(int pathLength) {
 		this.pathLength = pathLength;
 	}
 
+	/**
+	 * @return Whether the fish is moving forward along its path or not.
+	 */
 	public boolean isMovingForward() {
 		return movingForward;
 	}
 
+	/**
+	 * @param movingForward The moving status we want to set the fish to.
+	 */
 	public void setMovingForward(boolean movingForward) {
 		this.movingForward = movingForward;
 	}
 
-	public void setAvgSize(int AvgSize) {
+	/**
+	 * @param AvgSize The size (in feet) the fish will be set to.
+	 */
+	public void setAvgSize(double AvgSize) {
 		this.avgSize = AvgSize;
 	}
 
+	/**
+	 * @return The average size of the fish (in feet)
+	 */
 	public double getAvgSize() {
 		return avgSize;
 	}
 
+	/**
+	 * @param imagePath The photo of the fish to be displayed in the notead.
+	 */
 	public void setRealPic(String imagePath) {
 		this.realPic = imagePath;
 	}
 
+	/**
+	 * @return The photo of the fish displayed in the notepad
+	 */
 	public String getRealPic() {
 		return realPic;
 	}
